clean-all: reorganize repo

+++ contracts/NFTLock.sol
@@ -8,25 +8,28 @@ import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
 import "./INFTLock.sol";
 
 contract NFTLock is INFTLock, IERC721Receiver {
-
     struct NFT {
-        uint externalTokenId;
+        uint256 externalTokenId;
         address tokenContractAddress;
         address depositor;
     }
 
-    mapping (uint256 => NFT) lockedTokenById;
+    mapping(uint256 => NFT) lockedTokenById;
 
     uint256[] releasedIds;
 
     uint256 totalLockedTokens;
 
     /** @dev Events
-    *
-    */
+     *
+     */
     event TokenLocked(uint256 tokenId, address from);
     event TokenUnlocked(uint256 tokenId, address unlockerAddress);
-    event NFTReceived(address contractAddress, uint256 externalTokenId, address from);
+    event NFTReceived(
+        address contractAddress,
+        uint256 externalTokenId,
+        address from
+    );
     event NFTWithdrawn(uint256 tokenId, address to);
 
     function onERC721Received(
@@ -34,8 +37,7 @@ contract NFTLock is INFTLock, IERC721Receiver {
         address from,
         uint256 tokenId,
         bytes calldata
-    ) external override returns(bytes4) {
-
+    ) external override returns (bytes4) {
         emit NFTReceived(msg.sender, tokenId, from);
 
         NFT memory token = NFT(tokenId, msg.sender, from);
@@ -45,7 +47,6 @@ contract NFTLock is INFTLock, IERC721Receiver {
     }
 
     function lockToken(NFT memory token) internal {
-
         // find a released id or take a new one
         uint256 internalTokenId;
         if (releasedIds.length > 0) {
@@ -62,7 +63,6 @@ contract NFTLock is INFTLock, IERC721Receiver {
     }
 
     function unlockToken(uint256 internalTokenId) internal {
-
         releasedIds.push(internalTokenId);
         delete lockedTokenById[internalTokenId];
         totalLockedTokens--;
@@ -71,22 +71,27 @@ contract NFTLock is INFTLock, IERC721Receiver {
     }
 
     function withdraw(uint256 internalTokenId) external override {
-
         NFT storage token = lockedTokenById[internalTokenId];
         address sendTo = token.depositor;
 
         if (isMintedOnTz(internalTokenId)) {
-
             address newNFTOwner;
             bool isBurned;
 
             (isBurned, newNFTOwner) = isBurnedOnTz(internalTokenId);
-            require(isBurned, "NFT Locker: Burn your wrapped token before to unlock");
+            require(
+                isBurned,
+                "NFT Locker: Burn your wrapped token before to unlock"
+            );
             sendTo = newNFTOwner;
         }
 
         uint256 externalTokenId = token.externalTokenId;
-        ERC721(token.tokenContractAddress).safeTransferFrom(address(this), sendTo, externalTokenId);
+        ERC721(token.tokenContractAddress).safeTransferFrom(
+            address(this),
+            sendTo,
+            externalTokenId
+        );
 
         // delete NFT from the storage
         unlockToken(internalTokenId);
@@ -95,13 +100,21 @@ contract NFTLock is INFTLock, IERC721Receiver {
     }
 
     //it's a mock of a call to oracle function
-    function isMintedOnTz(uint256 internalTokenId) internal pure returns (bool) {
+    function isMintedOnTz(uint256 internalTokenId)
+        internal
+        pure
+        returns (bool)
+    {
         internalTokenId++;
         return true;
     }
 
     //it's a mock of a call to oracle function
-    function isBurnedOnTz(uint256 internalTokenId) internal pure returns (bool, address) {
+    function isBurnedOnTz(uint256 internalTokenId)
+        internal
+        pure
+        returns (bool, address)
+    {
         internalTokenId++;
         return (true, address(0));
     }
